requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.26.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
pdfplumber==0.10.3
Pillow==10.2.0
pillow-heif==0.21.0
python-dotenv==1.0.1
openai==1.65.5
PyPDF2==3.0.1 
